backend/requirements.txt
fastapi
python-dotenv
discord.py
